Removed Edit and Delete Buttons
I looked at the specs and it doesn't look like they want users to be able to change the database with edit and delete buttons. i think it's a fair assumption that this will be done by people operating directly on the database and not by users of this software. Therefore only the add, import, export and assign jobs buttons are left as per the specs.